Move the loading of @conf to increase flexibility
It would be nice to be able to change certain items that are currently
hard-coded into irc-bot.rb. As an example, I wanted to run Mathetes on
OFTC rather than Freenode.

In order to make that work, I moved the initial loading of @conf into
initialize - easy enough. However, since initialize itself calls reset
just a few lines later (which had been loading @conf), I also added a
param to reset to control whether or not the config is loaded. By default
the config will be loaded when reset is called, but you can simply change
that by calling reset(false).

I also placed a default value into the mathetes-config.yaml.sample, so
that Freenode would continue to be the default server.